// apps/frontend/src/react/api/generated/models/ImageTextBoundingBoxDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ImageTextBoundingBoxDto = {
    /**
     * 번역 대상 텍스트
     */
    text: string;
    /**
     * 이미지 내 텍스트의 위치를 나타내는 [y1, x1, y2, x2] 좌표
     */
    box_2d: Array<number>;
};